requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
idna==3.2
mouse==0.7.1
MouseInfo==0.1.3
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dotenv==0.19.0
PyTweening==1.0.3
requests==2.26.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.6
webdriver-manager==3.4.2
